Make file importing responsibility of import view (selectfile local)
